basic export cli
