+++ crates/editor/src/lib.rs
@@ -2,4 +2,4 @@ mod editor;
 mod editor_instance;
 mod playback;
 
-pub use editor_instance::{EditorInstance, EditorState, Segment};
+pub use editor_instance::{create_segments, EditorInstance, EditorState, Segment};
